refactors extract
